Merge branch 'feature/MKA-681-ios-auto-deploy' into 'develop'
feature(MKA-681): Switched to public version of rn-background-geolocation

There is no easy way to access private repo of this library so we decided
to follow public version (which is up to date now).

Closes MKA-681

feature(MKA-681): Switched to public version of rn-background-geolocation

There is no easy way to access private repo of this library so we decided
to follow public version (which is up to date now).

feature(MKA-681): Switched to public version of rn-background-geolocation

There is no easy way to access private repo of this library so we decided
to follow public version (which is up to date now).